refactor: Rename files to follow correct naming conventions
