--- Week0v2/Engine/Source/Runtime/Engine/Classes/Engine/Assets/Animation/AnimDataModel.cpp
@@ -1,17 +1,0 @@
-#include "AnimDataModel.h"
-
-UAnimDataModel::UAnimDataModel()
-    : PlayLength(0)
-    , NumberOfFrames(0)
-    , NumberOfKeys(0)
-{
-}
-
-UAnimDataModel::~UAnimDataModel()
-{
-}
-
-const TArray<FBoneAnimationTrack>& UAnimDataModel::GetBoneAnimationTracks() const
-{
-    return BoneAnimationTracks;
-}
