--- Week0v2/Engine/Source/Runtime/Engine/Classes/Engine/Assets/Animation/AnimTypes.cpp
@@ -1,1 +1,0 @@
-#include "AnimTypes.h"
