--- Week0v2/Engine/Source/Runtime/Engine/Classes/Engine/Assets/UAsset.cpp
@@ -1,5 +1,0 @@
-#include "UAsset.h"
-
-UAsset::~UAsset()
-{
-}
